Add files via upload
Course Resources: Full-Stack Web Development Bootcamp
In the "Course Resources" folder, you will find all the questions and materials for each submodule. Each submodule includes:

Goal Image: A visual representation of what needs to be built in that section.
Index: The data and specifications that need to be worked on.
Solution: The answers and solutions provided by the course.
Additionally, the HTML files excluding those in the "Solution" folder are my own implementations and solutions for the "Introduction to HTML" module, where I applied the concepts learned to build projects and practice HTML coding.